Big commit
Committing the updated version of the apex finding code. This has some small changes to the logic to determine what is or isn't an apex as well as introducing the function to save the results for each circuit. Also have developed the code at the end in terms of presenting the overall results for all the circuits. Also wrote a new script to create an animated figure for the powerpoint showing the category of each corner for each circuit. I've also ended up adding some tyre plot code to this branch, it shouldnt really be here but I can't be bothered to move it now